Move some function definitions/prototypes to solve inconsistent dll linkage error.
