+++ src/common/mmgcommon.h
@@ -519,24 +519,6 @@ void MMG5_excfun(int sigid) {
   void nl##__ pl                                  \
   { body }                                        \
 
-/**
- * \enum MMG5_Format
- * \brief Type of supported file format
- */
-enum MMG5_Format {
-  MMG5_FMT_MeditASCII, /*!< ASCII Medit (.mesh) */
-  MMG5_FMT_MeditBinary, /*!< Binary Medit (.meshb) */
-  MMG5_FMT_GmshASCII, /*!< ASCII Gmsh */
-  MMG5_FMT_GmshBinary, /*!< Binary Gmsh */
-  MMG5_FMT_VtkPvtp, /*!< VTK pvtp */
-  MMG5_FMT_VtkPvtu, /*!< VTK pvtu */
-  MMG5_FMT_VtkVtu, /*!< VTK vtu */
-  MMG5_FMT_VtkVtp, /*!< VTK vtp */
-  MMG5_FMT_VtkVtk, /*!< VTK vtk */
-  MMG5_FMT_Tetgen, /*!< Tetgen or Triangle */
-  MMG5_FMT_Unknown /*!< Unrecognized */
-};
-
 
 /* Global variables */
   static const uint8_t MMG5_inxt2[6] = {1,2,0,1,2}; /*!< next vertex of triangle: {1,2,0} */
@@ -584,7 +566,7 @@ typedef struct MMG5_dNode_s {
  extern double MMG5_det4pt(double c0[3],double c1[3],double c2[3],double c3[3]);
  int           MMG5_devangle(double* n1, double *n2, double crit);
  extern double MMG5_orvol(MMG5_pPoint point,int *v);
- LIBMMG_CORE_EXPORT int           MMG5_Add_inode( MMG5_pMesh mesh, MMG5_iNode **liLi, int val );
+ int           MMG5_Add_inode( MMG5_pMesh mesh, MMG5_iNode **liLi, int val );
  int           MMG5_Add_dnode( MMG5_pMesh mesh, MMG5_dNode **liLi, int, double);
  int           MMG5_eigenvmatsym2d(MMG5_pMesh mesh,double m[],double lambda[],double v[][2]);
  int           MMG5_eigenvmatsym3d(MMG5_pMesh mesh,double m[],double lambda[],double v[][3]);
@@ -594,7 +576,7 @@ typedef struct MMG5_dNode_s {
  int           MMG5_buildridmet(MMG5_pMesh,MMG5_pSol,int,double,double,double,double*,double[3][3]);
  extern int    MMG5_buildridmetfic(MMG5_pMesh,double*,double*,double,double,double,double*);
  int           MMG5_buildridmetnor(MMG5_pMesh, MMG5_pSol, int,double*, double*,double[3][3]);
-void           MMG5_check_hminhmax(MMG5_pMesh mesh, int8_t sethmin, int8_t sethmax);
+ void           MMG5_check_hminhmax(MMG5_pMesh mesh, int8_t sethmin, int8_t sethmax);
  int           MMG5_paratmet(double c0[3],double n0[3],double m[6],double c1[3],double n1[3],double mt[6]);
  void          MMG5_transpose3d(double m[3][3]);
  void          MMG5_dotprod(int8_t dim,double *a,double *b,double *result);
@@ -620,16 +602,10 @@ void           MMG5_check_hminhmax(MMG5_pMesh mesh, int8_t sethmin, int8_t sethm
                                    MMG5_Bezier*,double r[3][3],double gv[2]);
  void          MMG5_fillDefmetregSys( int, MMG5_pPoint, int, MMG5_Bezier,double r[3][3],
                                        double *, double *, double *, double *);
- LIBMMG_CORE_EXPORT void          MMG5_Free_ilinkedList( MMG5_pMesh mesh, MMG5_iNode *liLi );
+ void          MMG5_Free_ilinkedList( MMG5_pMesh mesh, MMG5_iNode *liLi );
  void          MMG5_Free_dlinkedList( MMG5_pMesh mesh, MMG5_dNode *liLi );
  LIBMMG_CORE_EXPORT int           MMG5_grad2metSurf(MMG5_pMesh,MMG5_pSol,MMG5_pTria,int,int);
  LIBMMG_CORE_EXPORT int           MMG5_grad2metSurfreq(MMG5_pMesh,MMG5_pSol,MMG5_pTria,int,int);
- LIBMMG_CORE_EXPORT char         *MMG5_Get_filenameExt( char *filename );
- char         *MMG5_Get_basename(char *path);
- char         *MMG5_Get_path(char *path);
- char         *MMG5_Remove_ext( char *path, char* );
- LIBMMG_CORE_EXPORT const char    *MMG5_Get_formatName(enum MMG5_Format fmt);
- LIBMMG_CORE_EXPORT int           MMG5_Get_format( char *ptr, int );
  int           MMG5_hashFace(MMG5_pMesh,MMG5_Hash*,int,int,int,int);
  int           MMG5_hashEdge(MMG5_pMesh mesh,MMG5_Hash *hash,int a,int b,int k);
  int           MMG5_hashUpdate(MMG5_Hash *hash,int a,int b,int k);
@@ -645,12 +621,12 @@ void           MMG5_check_hminhmax(MMG5_pMesh mesh, int8_t sethmin, int8_t sethm
  void          MMG5_mmgDefaultValues(MMG5_pMesh mesh);
  int           MMG5_mmgHashTria(MMG5_pMesh mesh, int *adja, MMG5_Hash*, int chkISO);
  void          MMG5_mmgInit_parameters(MMG5_pMesh mesh);
- LIBMMG_CORE_EXPORT void          MMG5_mmgUsage(char *prog);
- LIBMMG_CORE_EXPORT void          MMG5_paramUsage1(void);
- LIBMMG_CORE_EXPORT void          MMG5_paramUsage2(void);
- LIBMMG_CORE_EXPORT void          MMG5_2d3dUsage(void);
- LIBMMG_CORE_EXPORT void          MMG5_lagUsage(void);
- LIBMMG_CORE_EXPORT void          MMG5_advancedUsage(void);
+ void          MMG5_mmgUsage(char *prog);
+ void          MMG5_paramUsage1(void);
+ void          MMG5_paramUsage2(void);
+ void          MMG5_2d3dUsage(void);
+ void          MMG5_lagUsage(void);
+ void          MMG5_advancedUsage(void);
  extern int    MMG5_nonUnitNorPts(MMG5_pMesh,int,int,int,double*);
  extern double MMG5_nonorsurf(MMG5_pMesh mesh,MMG5_pTria pt);
  extern int    MMG5_norpts(MMG5_pMesh,int,int,int,double *);
@@ -678,19 +654,19 @@ void           MMG5_check_hminhmax(MMG5_pMesh mesh, int8_t sethmin, int8_t sethm
  double        MMG5_surftri33_ani(MMG5_pMesh,MMG5_pTria,double*,double*,double*);
  double        MMG5_surftri_iso(MMG5_pMesh mesh,MMG5_pSol met,MMG5_pTria ptt);
  extern int    MMG5_sys33sym(double a[6], double b[3], double r[3]);
- LIBMMG_CORE_EXPORT int           MMG5_unscaleMesh(MMG5_pMesh mesh,MMG5_pSol met,MMG5_pSol ls);
+ int           MMG5_unscaleMesh(MMG5_pMesh mesh,MMG5_pSol met,MMG5_pSol ls);
  int           MMG5_interpreg_ani(MMG5_pMesh,MMG5_pSol,MMG5_pTria,int8_t,double,double *mr);
  int           MMG5_interp_iso(double *ma,double *mb,double *mp,double t);
  int           MMG5_intersecmet22(MMG5_pMesh mesh, double *m,double *n,double *mr);
- LIBMMG_CORE_EXPORT extern int    MMG5_countLocalParamAtTri( MMG5_pMesh,MMG5_iNode **);
- LIBMMG_CORE_EXPORT extern int    MMG5_writeLocalParamAtTri( MMG5_pMesh,MMG5_iNode *,FILE*);
- double         MMG2D_quickarea(double a[2],double b[2],double c[2]);
+ int           MMG5_countLocalParamAtTri( MMG5_pMesh,MMG5_iNode **);
+ int           MMG5_writeLocalParamAtTri( MMG5_pMesh,MMG5_iNode *,FILE*);
+ double        MMG2D_quickarea(double a[2],double b[2],double c[2]);
  void          MMG5_build3DMetric(MMG5_pMesh mesh,MMG5_pSol sol,int ip,double dbuf[6]);
  int           MMG5_loadVtuMesh(MMG5_pMesh mesh,MMG5_pSol sol,const char *filename);
- int            MMG5_loadMshMesh_part1(MMG5_pMesh mesh,const char *filename,
-                                       FILE **inm,long *posNodes, long *posElts,
-                                       long **posNodeData, int *bin, int *iswp,
-                                       int *nelts,int *nsols);
+ int           MMG5_loadMshMesh_part1(MMG5_pMesh mesh,const char *filename,
+                                      FILE **inm,long *posNodes, long *posElts,
+                                      long **posNodeData, int *bin, int *iswp,
+                                      int *nelts,int *nsols);
  int            MMG5_check_readedMesh(MMG5_pMesh mesh,int nref);
  int            MMG5_loadMshMesh_part2(MMG5_pMesh mesh,MMG5_pSol *sol,FILE **inm,
                                        const long posNodes,const long posElts,
@@ -782,10 +758,6 @@ void MMG5_keep_subdomainElts ( MMG5_pMesh,int,int (*delElt)(MMG5_pMesh,int) );
 /* useful functions to debug */
  LIBMMG_CORE_EXPORT extern int  (*MMG5_indElt)(MMG5_pMesh mesh,int kel);
  LIBMMG_CORE_EXPORT extern int  (*MMG5_indPt)(MMG5_pMesh mesh,int kp);
-
-#ifdef USE_SCOTCH
-extern int    (*MMG5_renumbering)(int,MMG5_pMesh,MMG5_pSol sol,MMG5_pSol fields,int*);
-#endif
 
 void   MMG5_Set_commonFunc(void);
