+++ src/common/tools.c
@@ -23,7 +23,7 @@
 
 /**
  * \file common/tools.c
- * \brief Various tools for the mmg applications.
+ * \brief Various tools for the mmg libraries.
  * \author Charles Dapogny (UPMC)
  * \author Cécile Dobrzynski (Bx INP/Inria/UBordeaux)
  * \author Pascal Frey (UPMC)
@@ -902,198 +902,6 @@ void MMG5_memOption_memSet(MMG5_pMesh mesh) {
   return;
 }
 
-/**
- * \param mesh pointer toward the mesh structure (for count of used memory).
- * \param node pointer toward a MMG5_iNode (cell for linked list)
- * \return 1 if we can alloc the node \a node, 0 otherwise.
- *
- * Node allocation.
- *
- */
-static inline
-int MMG5_Alloc_inode( MMG5_pMesh mesh, MMG5_iNode **node ) {
-
-  MMG5_ADD_MEM(mesh,sizeof(MMG5_iNode),"boundary reference node",
-                return 0;);
-
-  MMG5_SAFE_MALLOC(*node,1,MMG5_iNode,return 0);
-
-  return 1;
-}
-
-/**
- * \param mesh pointer toward the mesh structure (for count of used memory).
- * \param liLi pointer toward the address of the root of the linked list.
- * \param val value to add to the linked list.
- * \return 1 if the node is inserted, 0 if the node is not inserted, -1 if fail.
- *
- * Add a node with value \a val to a sorted linked list with unique entries.
- *
- * \remark as the linked list had unique entries, we don't insert a node if it
- * exists.
- *
- */
-int MMG5_Add_inode( MMG5_pMesh mesh, MMG5_iNode **liLi, int val ) {
-  MMG5_iNode  *newNode, *cur;
-
-  cur = *liLi;
-
-  /* Travel through the linked list and search if the value val exist or, if
-   * not, where to insert it */
-  if ( cur ) {
-    if ( val < (*liLi)->val ) {
-      /* Add a value at the list head */
-      if ( !MMG5_Alloc_inode(mesh,&newNode) ) return -1;
-
-      newNode->val = val;
-      newNode->nxt = (*liLi);
-
-      (*liLi) = newNode;
-
-      return 1;
-
-    }
-    else if (val == (*liLi)->val ) return 0;
-
-    while ( cur->nxt && ( val >= (cur->nxt)->val) )
-      cur = cur->nxt;
-
-    if ( val == cur->val ) return 0;
-
-    if ( !MMG5_Alloc_inode(mesh,&newNode) ) return -1;
-
-    newNode->val = val;
-    newNode->nxt = cur->nxt;
-    cur->nxt = newNode;
-  }
-  else {
-    if ( !MMG5_Alloc_inode(mesh,&newNode) ) return -1;
-
-    newNode->val = val;
-    newNode->nxt = NULL;
-
-    *liLi = newNode;
-  }
-
-  return 1;
-}
-
-/**
- * \param mesh pointer toward the mesh structure (for count of used memory).
- * \param liLi pointer toward the root of the linked list.
- *
- * Free the memory used by the linked list whose root is \a liLi.
- *
- */
-void MMG5_Free_ilinkedList( MMG5_pMesh mesh, MMG5_iNode *liLi ) {
-  MMG5_iNode *cur,*nxt;
-
-  cur = liLi;
-  while (cur) {
-    nxt = cur;
-    cur = cur->nxt;
-
-    MMG5_DEL_MEM(mesh,nxt);
-  }
-}
-
-
-/**
- * \param mesh pointer toward the mesh structure (for count of used memory).
- * \param node pointer toward a MMG5_dNode (cell for linked list)
- * \return 1 if we can alloc the node \a node, 0 otherwise.
- *
- * Node allocation.
- *
- */
-static inline
-int MMG5_Alloc_dnode( MMG5_pMesh mesh, MMG5_dNode **node ) {
-
-  MMG5_ADD_MEM(mesh,sizeof(MMG5_dNode),"node for hausdorff eval",
-                return 0;);
-
-  MMG5_SAFE_MALLOC(*node,1,MMG5_dNode,return 0);
-
-  return 1;
-}
-
-/**
- * \param mesh pointer toward the mesh structure (for count of used memory).
- * \param liLi pointer toward the address of the root of the linked list.
- * \param k integer value to add to the linked list.
- * \param val real value to add to the linked list.
- * \return 1 if the node is inserted, 0 if the node is not inserted, -1 if fail.
- *
- * Add a node with integer value \a k and real value \a val to a sorted linked
- * list with unique entries.
- *
- * \remark as the linked list had unique entries, we don't insert a node if it
- * exists.
- *
- */
-int MMG5_Add_dnode( MMG5_pMesh mesh, MMG5_dNode **liLi, int k, double val ) {
-  MMG5_dNode  *newNode, *cur;
-
-  cur = *liLi;
-
-  /* Travel through the linked list and search if the value val exist or, if
-   * not, where to insert it */
-  if ( cur ) {
-    if ( val < (*liLi)->val ) {
-      /* Add a value at the list head */
-      if ( !MMG5_Alloc_dnode(mesh,&newNode) ) return -1;
-
-      newNode->val = val;
-      newNode->nxt = (*liLi);
-
-      (*liLi) = newNode;
-
-      return 1;
-
-    }
-    else if (val == (*liLi)->val ) return 0;
-
-    while ( cur->nxt && ( val >= (cur->nxt)->val) )
-      cur = cur->nxt;
-
-    if ( val == cur->val ) return 0;
-
-    if ( !MMG5_Alloc_dnode(mesh,&newNode) ) return -1;
-
-    newNode->val = val;
-    newNode->nxt = cur->nxt;
-    cur->nxt = newNode;
-  }
-  else {
-    if ( !MMG5_Alloc_dnode(mesh,&newNode) ) return -1;
-
-    newNode->val = val;
-    newNode->nxt = NULL;
-
-    *liLi = newNode;
-  }
-
-  return 1;
-}
-
-/**
- * \param mesh pointer toward the mesh structure (for count of used memory).
- * \param liLi pointer toward the root of the linked list.
- *
- * Free the memory used by the linked list whose root is \a liLi.
- *
- */
-void MMG5_Free_dlinkedList( MMG5_pMesh mesh, MMG5_dNode *liLi ) {
-  MMG5_dNode *cur,*nxt;
-
-  cur = liLi;
-  while (cur) {
-    nxt = cur;
-    cur = cur->nxt;
-
-    MMG5_DEL_MEM(mesh,nxt);
-  }
-}
 
 /** Compute 3 * 3 determinant : det(c1-c0,c2-c0,v) */
 inline double MMG5_det3pt1vec(double c0[3],double c1[3],double c2[3],double v[3]) {
