+++ src/mmg3d/libmmg3d_tools.c
@@ -31,6 +31,7 @@
  **/
 
 #include "libmmg3d.h"
+#include "mmgcommon.h"
 #include "inlined_functions_3d.h"
 #include "mmgversion.h"
