+++ src/mmg3d/mmg3d.c
@@ -58,7 +58,6 @@ static void MMG5_endcod(void) {
  * references.
  *
  */
-static inline
 int MMG5_countLocalParamAtTet( MMG5_pMesh mesh,MMG5_iNode **bdyRefs) {
   int         npar,k,ier;
 
@@ -102,7 +101,6 @@ int MMG5_countLocalParamAtTet( MMG5_pMesh mesh,MMG5_iNode **bdyRefs) {
  * Write the local default values at tetrahedra in the parameter file.
  *
  */
-static inline
 int MMG5_writeLocalParamAtTet( MMG5_pMesh mesh, MMG5_iNode *bdryRefs,
                                 FILE *out ) {
   MMG5_iNode *cur;
@@ -128,7 +126,6 @@ int MMG5_writeLocalParamAtTet( MMG5_pMesh mesh, MMG5_iNode *bdryRefs,
  * can be locally defined.
  *
  */
-static inline
 int MMG3D_writeLocalParam( MMG5_pMesh mesh ) {
   MMG5_iNode  *triRefs,*tetRefs;
   int          nparTri,nparTet;
